Move utils.ts to utils folder
