+++ core/generator/db/dbGenerator.ts
@@ -1,7 +1,7 @@
 // core/generator/db/dbGenerator.ts
 import { ModelMeta } from '../../types/ModelMeta';
 import { generateDrizzleModel } from './drizzleAdapter';
-import { writeToFile } from '../utils';
+import { writeToFile } from '../utils/utils';
 import path from 'path';
 
 export async function generateDbModels(models: ModelMeta[]) {
